% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\name{as}
\alias{as}
\title{as("list", "cabinet")}
\description{
as() reads the data structure as used in the
donordata package, and transforms it into a cabinet object.
}
\seealso{
Other cabinet: \code{\link{list2cabinet}}
}

