% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bse.R
\docType{class}
\name{bse-class}
\alias{bse-class}
\title{Class for broken stick estimates}
\description{
The \code{bse} class stores three variables that result from predicting
a given broken stick model. The \code{bse} class extend
the \code{\link{xyz-class}} class. The \code{bse} class calculates
predicted values according to the broken stick model applied to the
values specified in the \code{data} argument.
}
\section{Slots}{

 \describe{
   \item{\code{call}:}{An object of class \code{call} that refers to an object with a pre-fitted broken stick model. Evaluation by \code{eval(call)} should yield an object of class \code{brokenstick.export} as created by
   \code{\link[brokenstick]{export.brokenstick}}.}
   \item{\code{zscale}:}{A logical indicating whether the broken stick model models the outcome in the Z-scale (\code{zscale = TRUE}, default) or Y-scale (\code{zscale = FALSE}). At initialization, the \code{new()} function tries to infer and set this flag from the model specified by \code{call}.}
}
}
\examples{
# specify three height measures
d <- new("bse", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4))
# Note that Z-scores are added, relative to nl1997 reference for males
d
# Same, but now for a female
d <- new("bse", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4), sex = "female")
d
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{xyz-class}}, \code{\link[brokenstick]{export.brokenstick}}
}
\keyword{classes}

