% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R, R/class-cabinet.R
\docType{class}
\name{individual-class}
\alias{cabinet-class}
\alias{individual-class}
\title{An S4 class to represent individual data}
\description{
A collection of object of \code{individualID}, \code{individualBG}, \code{individualAN} and \code{individualBS}, representing data of an individual.
The type of information covers individual identifyers, fixed background variables, time-varying anthroponetric measures and time-varying broken stick estimates. The \code{new("individual")} function can automatically calculate standard deviation scores relative to a growth reference, and predictions according the a broken stick model.

A collection of objects of class \code{individual} can be stored
into the class \code{cabinet} class.
The \code{cabinet} object can be generated by the \code{new("cabinet", ...)} and
\code{as(data, "cabinet")} functions.
The result is essentially a list of \code{individual} objects with an index.
Data of individual persons can be assessed by position by as list index
\code{`[[`} or by unique id by the vector index \code{`[`}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of objects of class \code{individual}}

\item{\code{ids}}{A number vector the indexes the (unique) individual \code{id}}

\item{\code{src}}{Name of the source data}

\item{\code{created}}{Creation date}

\item{\code{updated}}{Update date}

\item{\code{.Data}}{A list of objects of class \code{individual}}

\item{\code{ids}}{A number vector the indexes the (unique) individual \code{id}}

\item{\code{src}}{Name of the source data}

\item{\code{created}}{Creation date}

\item{\code{updated}}{Update date}
}}
\examples{
\dontrun{
# convert all 1933 children of SMOCC in `individual` objects
library("donordata")
all <- as(smocc, "cabinet")

# same, but not relative to WHO references
all2 <- list2cabinet(smocc, libname = "who", prefix = "who2011", sub = "")
}
\dontrun{
# convert all 1933 children of SMOCC in `individual` objects
library("donordata")
all <- as(smocc, "cabinet")

# same, but not relative to WHO references
all2 <- list2cabinet(smocc, libname = "who", prefix = "who2011", sub = "")
}
}
\author{
Stef van Buuren 2016

Stef van Buuren 2016
}
\seealso{
\code{\link[=individual-class]{individual}}

\code{\link[=individual-class]{individual}}
}
\keyword{classes}

