% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bds_individual.R
\name{convert_bds_individual}
\alias{convert_bds_individual}
\title{Convert json BSD data for single individual to class individual}
\usage{
convert_bds_individual(txt = NULL, ...)
}
\arguments{
\item{txt}{a JSON string, URL or file}

\item{\dots}{Additional parameter passed down to
\code{fromJSON(txt, ...)}, \code{new("xyz",... )} and
\code{new("bse",... )}. Useful parameters are \code{models =
"bsmodel"} for setting the broken stick model, or \code{call =
as.call(...)} for setting proper reference standards.}
}
\value{
An object of class \linkS4class{individual}.
}
\description{
This function takes data from a json source and saves as a an object
of class \linkS4class{individual}. The function automatically calculates
standard deviation scores and broken stick conditional means per visit.
}
\examples{
library(donorloader)
smocc_bs <- load_data(dnr = "smocc_bs")
fn <- file.path(path.package("minihealth"), "testdata", "client3.json")
p <- convert_bds_individual(fn)
}
\seealso{
\linkS4class{individual}, \linkS4class{bse}, \linkS4class{xyz},
         \code{\link[jsonlite]{fromJSON}}
}
\author{
Stef van Buuren 2019
}
