% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_individual_bds.R
\name{convert_individual_bds}
\alias{convert_individual_bds}
\title{Convert data of class individual to bds format}
\usage{
convert_individual_bds(ind = NULL, ...)
}
\arguments{
\item{ind}{Object of class \code{individual}}

\item{\dots}{Additional parameters. Currently ignored.}
}
\value{
Data in bds format as JSON
}
\description{
This function takes data from an object of class
\linkS4class{individual} and saves it into JSON bds format (which
can be sent to JAMES).
}
\details{
Functions \code{convert_bds_individual()} and
\code{convert_individual_bds()} are inverse operations.
}
\examples{
data("installed.cabinets", package = "jamestest")
ind <- installed.cabinets[[2]][[1]]
b <- convert_individual_bds(ind)
}
\seealso{
\linkS4class{individual}, \linkS4class{bse}, \linkS4class{xyz},
         \code{\link[jsonlite]{toJSON}}
}
\author{
Stef van Buuren 2019
}
