% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.donordata.R
\name{donordata.to.person}
\alias{donordata.to.person}
\title{Convert single person donor data to person class}
\usage{
donordata.to.person(id, src = "smocc", ...)
}
\arguments{
\item{id}{the id number of the person in specified source}

\item{src}{character vector with the name of the dataset. Default is \code{src = "smocc"}}

\item{\dots}{Additional parameter passed down to \code{new("xyz",... )} and \code{new("bse",... )}. Useful parameters are \code{model = "bsmodel"} for setting the broken stick model, or \code{call = as.call(...)} for setting proper reference standards.}
}
\description{
This function takes data from the \pkg{donordata} package, extract cases identified by \code{id} and save as a an object of class \code{person}. The function automatically calculates standard deviation scores and broken stick conditional means per visit.
}
\examples{
require("donordata")
p <- donordata.to.person(10001)
p
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}

