% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get a range from the antropometric data}
\usage{
get_range(ind, ynames = c("hgt", "wgt", "hdc", "bmi"), slotname = "x")
}
\arguments{
\item{ind}{An object of class \code{individual}.}

\item{ynames}{A vector of name of outcome variable}

\item{slotname}{The name of the slot, either \code{"x"} (often age),
\code{"y"} (measure) or \code{"z"} (SDS).}
}
\value{
A numeric vector of length 2 (if success), or
\code{numeric(0)} (if failed)
}
\description{
Get a range from the antropometric data
}
\examples{
# Find age range of anthropometric measures
ind <- new("individual",
           hgt = new("xyz", x = c(0.3, 2, 6)),
           wgt = new("xyz", x = c(0.2, 1), y = c(1, NA)))
get_range(ind)
get_range(ind, slotname = "y")
get_range(ind, ynames = c("hdc", "bmi"))
}
