% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz.R
\name{get_xyz}
\alias{get_xyz}
\title{Get the x, y and z coordinates of the measurement}
\usage{
get_xyz(ind, yname)
}
\arguments{
\item{ind}{An object of class \code{individual}.}

\item{yname}{A vector of name of outcome variable}
}
\value{
A data frame
}
\description{
Get the x, y and z coordinates of the measurement
}
\examples{
# Extract anthropometric measures
ind <- new("individual",
           hgt = new("xyz", x = c(0.3, 2, 6)),
           wgt = new("xyz", x = c(0.2, 1), y = c(1, NA)))
get_xyz(ind, "wgt")
get_xyz(ind, "hgt")
}
