% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\docType{class}
\name{individual-class}
\alias{individual-class}
\title{An S4 class to represent individual data}
\description{
A collection of object of \code{individualID}, \code{individualBG}, \code{individualAN} and \code{individualBS}, representing data of an individual.
The type of information covers individual identifyers, fixed background variables, time-varying anthroponetric measures and time-varying broken stick estimates. The \code{new("individual")} function can automatically calculate standard deviation scores relative to a growth reference, and predictions according the a broken stick model.
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link[=individualID-class]{individualID}}
}
\keyword{classes}

