% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.to.donordata.R
\name{individual.to.donordata}
\alias{individual.to.donordata}
\title{Convert object of class 'individual' to donordata}
\usage{
individual.to.donordata(x, type = NULL)
}
\arguments{
\item{x}{An object of class \code{individual}.}

\item{type}{The type of slot. This one of \code{"id"}, \code{"child"} and
\code{"time"}. The default \code{type = NULL} return a list of all three slots.}
}
\value{
This function returns slots \code{"id"}, \code{"child"} and
\code{"time"}, or a list of these slots (if \code{type = NULL}).
}
\description{
Convert object of class 'individual' to donordata
}
\details{
The code assumes that slots \code{"id"}, \code{"child"} and
\code{"time"} are up-to-data.
}
\examples{
require("donordata")
p <- donordata.to.individual(src = "smocc", id = 10001)
p
q <- individual.to.donordata(p)
q
}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}
\author{
Stef van Buuren 2017
}
