% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\docType{class}
\name{individualAN-class}
\alias{individualAN-class}
\title{An S4 class to represent individual anthropometric data}
\description{
An S4 class to represent individual anthropometric data
}
\section{Slots}{

\describe{
\item{\code{hgt}}{Length/height in cm (\code{xyz})}

\item{\code{wgt}}{Weight in kg (\code{xyz})}

\item{\code{hdc}}{Head circumference in cm (\code{xyz})}
}}
\author{
Stef van Buuren 2016
}

