% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualBG.R
\docType{class}
\name{individualBG-class}
\alias{individualBG-class}
\title{An S4 class to represent individual background variables}
\description{
This class stores individual time-invariant background variables,
such as sex or birth weight. Each slot has length one.
}
\section{Slots}{

\describe{
\item{\code{sex}}{Sex: \code{"male"}, \code{"female"} (\code{character})}

\item{\code{etn}}{Etnicity: \code{"NL"}, \code{"TU"},
\code{"MA"}, \code{"HS"}, \code{"other"} (\code{character})}

\item{\code{edu}}{Education: \code{"low"}, \code{"middle"}, \code{"high"}}

\item{\code{gad}}{Gestational age, in days (\code{numeric})}

\item{\code{ga}}{Gestational age, in completed weeks (\code{numeric})}

\item{\code{bw}}{Birth weight, grammes (\code{numeric})}

\item{\code{twin}}{Twin: \code{0}, \code{1} (\code{numeric})}

\item{\code{agem}}{Mother age, years (\code{numeric})}

\item{\code{smo}}{Smoked during pregnancy, \code{0}, \code{1} (\code{numeric})}

\item{\code{hgtm}}{Height of mother in cm (\code{numeric})}

\item{\code{wgtm}}{Weight of mother in kg (\code{numeric})}

\item{\code{hgtf}}{Height of father in cm (\code{numeric})}

\item{\code{wgtf}}{Weight of father in kg (\code{numeric})}

\item{\code{bfexc06}}{Days exclusive breastfeeding, 0-6 m (\code{numeric})}

\item{\code{durbrst}}{Days mixed breastfeeding 0-6m (\code{numeric})}
}}

