% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualBS.R
\docType{class}
\name{individualBS-class}
\alias{individualBS-class}
\title{An S4 class to represent individual broken stick estimates}
\description{
An S4 class to represent individual broken stick estimates
}
\section{Slots}{

\describe{
\item{\code{hgt}}{Length/height in cm (\code{bse})}

\item{\code{wgt}}{Weight in kg (\code{bse})}

\item{\code{hdc}}{Head circumference in cm (\code{bse})}

\item{\code{bmi}}{Body mass index kg/m**2 (\code{bse})}

\item{\code{wfh}}{Weight for height kg/m (\code{bse})}

\item{\code{dsc}}{D-score (D) (\code{bse}) - currently unavailable}
}}

\author{
Stef van Buuren 2016-2020
}
