% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualDS.R
\docType{class}
\name{individualDS-class}
\alias{individualDS-class}
\title{An S4 class to represent individual D-score data}
\description{
The \code{individualDS} class stores individual milestone 0/1 data
and calculates the D-score and DAZ.
}
\section{Slots}{

\describe{
\item{\code{mst}}{An numeric \code{matrix} or a \code{data.frame} with
milestone responses data.
The matrix should contain a column indicating age of measurement
for each row, and columns with 0/1/NA responses on
milestones. Colomn names of milestones must adhere to the 9-position
GSED milestones naming convention.}

\item{\code{dfa}}{D-score for age in D-units (\code{xyz})}

\item{\code{xname}}{Name of the age variable}

\item{\code{xunit}}{Time unit of the age variable}

\item{\code{key}}{The conversion key for the D-score}

\item{\code{population}}{The reference population for DAZ calculation}
}}

\seealso{
\code{\link{xyz-class}}, \code{\link[dscore]{dscore}}
}
\author{
Stef van Buuren 2019
}
\keyword{classes}
