% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualID.R
\docType{class}
\name{individualID-class}
\alias{individualID-class}
\title{An S4 class to represent individual identifiers}
\description{
An S4 class to represent individual identifiers
}
\section{Slots}{

\describe{
\item{\code{id}}{Unique numerical individual identifier in source (\code{integer}, length 1)}

\item{\code{name}}{Name of the individual (\code{character})}

\item{\code{dob}}{Date of birth of class (\code{date}, length 1)}

\item{\code{src}}{Name of data source (\code{character})}
}}
\examples{
# Create a new ID for Ron and Jasper
ron <- new("individualID", name = c("Ron", "Smith"),
           dob = as.Date("1999-08-22"), id = as.integer(204))
jasper <- new("individualID", name = c("Jasper", "Fielding"),
           id = as.integer(220))
}
\author{
Stef van Buuren 2016
}

