% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\name{is.cabinet}
\alias{is.cabinet}
\title{Is this object of class `cabinet`?}
\usage{
is.cabinet(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical
}
\description{
Is this object of class `cabinet`?
}

