% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\name{is.individual}
\alias{is.individual}
\title{Is this object of class `individual`?}
\usage{
is.individual(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical
}
\description{
Is this object of class `individual`?
}
