% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\name{list2cabinet}
\alias{list2cabinet}
\title{Convert donordata object to cabinet object}
\usage{
list2cabinet(from, ...)
}
\arguments{
\item{from}{Longitudinal data in the \code{list} format as used by the \code{donordata} package}

\item{\dots}{Argument passed down to \code{donordata.to.individual}}
}
\description{
This function calls the \code{donordata.to.individual} function to
transform data of each \code{id} into an object of \code{individual} S4 class,
and then collects all individuals into an object of \code{cabinet} S4 class.
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{donordata.to.individual}}

Other cabinet: \code{\link{as}}
}

