% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\docType{methods}
\name{[,cabinet,ANY,ANY,ANY-method}
\alias{[,cabinet,ANY,ANY,ANY-method}
\title{Extract individuals from cabinet}
\usage{
\S4method{[}{cabinet,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An object of class \code{cabinet}}

\item{i}{Index vector}
}
\description{
Extract individuals from cabinet
}

