% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\name{[<-,cabinet,ANY,ANY,ANY-method}
\alias{[<-,cabinet,ANY,ANY,ANY-method}
\title{Replace part of cabinet}
\usage{
\S4method{[}{cabinet,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{An S4 object of class `cabinet`}

\item{i}{Indices specifying elements to replace}

\item{value}{An S4 object of class `cabinet` or `individual`}
}
\description{
Replace part of cabinet
}
