% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_individual.R
\name{tidy_individual}
\alias{tidy_individual}
\title{Tidy individual class}
\usage{
tidy_individual(ind)
}
\arguments{
\item{ind}{Object of class \code{individual}}

\item{\dots}{Additional parameters. Currently ignored.}
}
\value{
A \code{data.frame}
}
\description{
This function takes data from an object of class
\linkS4class{individual} and saves a \code{data.frame}
with anthropometric measures.
}
\examples{
data("installed.cabinets", package = "jamestest")
ind <- installed.cabinets[[3]][[8]]
b <- tidy_individual(ind)
}
\seealso{
\linkS4class{individual}
}
\author{
Stef van Buuren 2020
}
