% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bds_individual.R
\name{validate_bds_individual}
\alias{validate_bds_individual}
\title{Validate JSON file}
\usage{
validate_bds_individual(txt = NULL)
}
\arguments{
\item{txt}{a JSON string, URL or file to be compared to the schema.}
}
\value{
A \code{boolean} with an \code{error} attribute.
}
\description{
This function takes data from a JSON source and compares it to a JSON schema
object. A \code{boolean} is returned indicating whether or not the provided
JSON data matches the schema.
}
\author{
Arjan Huizing 2019
}
