% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xyz.R, R/class-bse.R, R/class-ird.R,
%   R/class-individualMS.R, R/class-individualAN.R, R/class-cabinet.R
\name{as}
\alias{as}
\title{as("xyz", "data.frame")}
\description{
as() reads the data structure as used in the
donordata package, and transforms it into a cabinet object.
}
\seealso{
Other cabinet: 
\code{\link{list2cabinet}()}
}
\concept{bse}
\concept{cabinet}
\concept{individualAN}
\concept{individualMS}
\concept{ird}
\concept{xyz}
