% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\docType{class}
\name{cabinet-class}
\alias{cabinet-class}
\title{Cabinet class}
\description{
A collection of objects of class \code{individual} can be stored
into the class \code{cabinet} class.
The \code{cabinet} object can be generated by the \code{new("cabinet", ...)} and
\code{as(data, "cabinet")} functions.
The result is essentially a list of \code{individual} objects with an index.
Data of individual persons can be assessed by position by as list index
\code{`[[`} or by unique id by the vector index \code{`[`}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of objects of class \code{individual}}

\item{\code{ids}}{A number vector the indexes the (unique) individual \code{id}}

\item{\code{n}}{Number of individuals}

\item{\code{readonly}}{Logical indicating whether the cabinet is just for viewing
(\code{TRUE}) or for editing by the user (\code{FALSE}, the default)}

\item{\code{created}}{Creation date}

\item{\code{updated}}{Update date}
}}
\examples{
# create cabinet with 1 empty record
z <- new("cabinet")
class(z)
slotNames(z)
slotNames(z[[1]])

# create cabinet with three empty records
z <- new("cabinet", n = 3)
slotNames(z)

# create cabinet with one individual, Sally
sally <- new("individual", name = "Sally", sex = "female", id = 22L)
z <- new("cabinet", sally)
length(z)

# create cabinet with two individuals, Harry and Sally
harry <- new("individual", name = "Harry", id = as.integer(33))
z <- new("cabinet", list(harry = harry, sally = sally))
z[["harry"]]@name; z[["harry"]]@sex
z[["sally"]]@name; z[["sally"]]@sex

\dontrun{
# convert all 1933 children of SMOCC in `individual` objects
library("donordata")
cab <- as(smocc, "cabinet")
smoccdemo <- cab[1:10]

# same, but not relative to WHO references
cab.who <- list2cabinet(smocc, libname = "who", prefix = "who2011", sub = "")
}
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link[=individual-class]{individual}}
}
\keyword{classes}

