% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\name{donordata.to.individual}
\alias{donordata.to.individual}
\title{Convert single individual from donor data to class individual}
\usage{
donordata.to.individual(id, src = donordata::smocc, ...)
}
\arguments{
\item{id}{the id number of the individual in specified source}

\item{src}{A character indicating the source}

\item{\dots}{Additional parameter passed down to \code{new("xyz",... )} and \code{new("bse",... )}. Useful parameters are \code{model = "bsmodel"} for setting the broken stick model, or \code{call = as.call(...)} for setting proper reference standards.}
}
\description{
This function takes data from the \pkg{donordata} package, extract cases identified by \code{id} and save as a an object of class \code{individual}. The function automatically calculates standard deviation scores and broken stick conditional means per visit.
}
\examples{
require("donordata")
p <- donordata.to.individual(10001)
p
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}

