% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\name{donordata.to.individual}
\alias{donordata.to.individual}
\title{Convert single individual from donor data to class individual}
\usage{
donordata.to.individual(id, src = donordata::smocc, ...)
}
\arguments{
\item{id}{the id number of the individual in specified source}

\item{src}{A character indicating the source}

\item{\dots}{Additional parameter passed down to \code{new("xyz",... )} and \code{new("bse",... )}. Useful parameters are \code{model = "bsmodel"} for setting the broken stick model, or \code{call = as.call(...)} for setting proper reference standards.}
}
\description{
This function takes data from the \pkg{donordata} package, extract cases identified by \code{id} and save as a an object of class \code{individual}. The function automatically calculates standard deviation scores and broken stick conditional means per visit.
}
\examples{
require("donordata")
p <- donordata.to.individual(10001)
p

# from lollypop.preterm
# calculating Z-score relative to preterm growth references
q <- donordata.to.individual(id = 50001, src = donordata::lollypop.preterm)

# overwrite hgt, wgt and hdc slots
q@hgt <- new("xyz", x = q@hgt@x, y = q@hgt@y, yname = "hgt",
             libname = "preterm", prefix = "pt2012a",
             sex = q@sex, sub = q@ga)
q@wgt <- new("xyz", x = q@wgt@x, y = q@wgt@y, yname = "wgt",
             libname = "preterm", prefix = "pt2012a",
             sex = q@sex, sub = q@ga)
q@hdc <- new("xyz", x = q@hdc@x, y = q@hdc@y, yname = "hdc",
             libname = "preterm", prefix = "pt2012b",
             sex = q@sex, sub = q@ga)

# update broken stick estimate
q@bs.hgt <- new("bse", data = q@hgt)
q@bs.wgt <- new("bse", data = q@wgt)
q@bs.hdc <- new("bse", data = q@hdc)

q

}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}
\author{
Stef van Buuren 2016
}
