% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donordata_to_individual.R
\name{donordata_to_individual}
\alias{donordata_to_individual}
\alias{donordata.to.individual}
\title{Convert single individual from donor data to class individual}
\usage{
donordata_to_individual(con = NULL, dnr, id, ...)

donordata.to.individual(con = NULL, dnr, id, ...)
}
\arguments{
\item{con}{A database connection. The default \code{con = NULL} reads the data from
the donordata package.}

\item{dnr}{A character indicating the source, e.g. \code{dnr = "smocc"}}

\item{id}{the id number of the individual in specified source. If specified as a vector,
only the first element is used.}

\item{\dots}{Additional parameter passed down to \code{new("xyz",... )} and \code{new("bse",... )}.
Useful parameters are \code{models = "bsmodel"} for setting the broken stick model,
or \code{call = as.call(...)} for setting proper reference standards.}
}
\value{
An object of class \code{individual}. If \code{id} is not found in data
\code{dnr}, then the function will set only slots \code{id} and \code{dnr}.
}
\description{
This function takes data from the \pkg{donordata} package, extract cases identified by \code{id} and save as a an object of class \code{individual}. The function automatically calculates standard deviation scores and broken stick conditional means per visit.
}
\note{
\code{donordata.to.individual()} is deprecated, but exported
for legacy reasons
}
\examples{
require("donorloader")
smocc <- load_data(dnr = "smocc")
smocc_bs <- load_data(dnr = "smocc_bs")
lollypop <- load_data(dnr = "lollypop")
terneuzen <- load_data(dnr = "terneuzen")
terneuzen_bs <- load_data(dnr = "terneuzen_bs")

p <- donordata_to_individual(dnr = "smocc", id = 10001)

# from lollypop
# calculating Z-score relative to preterm growth references
q <- donordata_to_individual(dnr = "lollypop", id = 50001)

# This is a manual way to convert to preterm references
# Not needed anymore after introduction of usetransform (0.81)
#q@hgt <- new("xyz", x = q@hgt@x, y = q@hgt@y, yname = "hgt",
#              libname = "clopus::preterm", prefix = "pt2012a",
#             sex = q@sex, sub = q@ga)
#q@wgt <- new("xyz", x = q@wgt@x, y = q@wgt@y, yname = "wgt",
#              libname = "clopus::preterm", prefix = "pt2012a",
#              sex = q@sex, sub = q@ga)
#q@hdc <- new("xyz", x = q@hdc@x, y = q@hdc@y, yname = "hdc",
#              libname = "clopus::preterm", prefix = "pt2012b",
#              sex = q@sex, sub = q@ga)
# # update broken stick estimate
# q@bs.hgt <- new("bse", data = q@hgt)
# q@bs.wgt <- new("bse", data = q@wgt)
# q@bs.hdc <- new("bse", data = q@hdc)

# use models argument to estimate brokenstick estimates
p <- donordata_to_individual(dnr = "terneuzen", id = 11,
  models = "terneuzen_bs")

}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}
\author{
Stef van Buuren 2017
}
