% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz.R
\name{get_xyz}
\alias{get_xyz}
\title{Get the x, y and z coordinates of the measurement}
\usage{
get_xyz(ind, yname)
}
\arguments{
\item{ind}{An object of class \code{individual}.}

\item{yname}{Name of outcome variable, e.g. \code{"hgt"}. Must be one
of \code{slotNames(ind)}.}
}
\value{
A data frame with names \code{"x"}, \code{"y"} and \code{"z"}
}
\description{
The function removes rows with incomplete \code{"x"} or \code{"y"}
components.
}
\examples{
# Extract anthropometric measures
ind <- new("individual",
           hgt = new("xyz", x = c(0.3, 2, 6)),
           wgt = new("xyz", x = c(0.2, 1), y = c(1, NA)))
get_xyz(ind, "wgt")
get_xyz(ind, "hgt")
}
