% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_to_donordata.R
\name{individual_to_donordata}
\alias{individual_to_donordata}
\title{Convert object of class 'individual' to donordata}
\usage{
individual_to_donordata(x, type = NULL)
}
\arguments{
\item{x}{An object of class \code{individual}.}

\item{type}{The type of slot. This one of \code{"id"}, \code{"child"} and
\code{"time"}. The default \code{type = NULL} return a list of all three slots.}
}
\value{
This function returns slots \code{"id"}, \code{"child"} and
\code{"time"}, or a list of these slots (if \code{type = NULL}).
}
\description{
This function returns the slots \code{"id"}, \code{"child"} and
\code{"time"}, either separate or as a \code{list} from
the \code{individual} object. The code
assumes that slots \code{"id"}, \code{"child"} and \code{"time"}
are up-to-data and consistent with the other slots.
}
\examples{
require("donordata")
p <- donordata_to_individual(dnr = "smocc", id = 10001)
p
q <- individual_to_donordata(p)
q
}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}
\author{
Stef van Buuren 2017
}
