% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_to_donordata.R
\name{individual_to_donordata}
\alias{individual_to_donordata}
\alias{individual.to.donordata}
\title{Convert object of class 'individual' to donordata}
\usage{
individual_to_donordata(x, elements = c("list", "child", "time"))

individual.to.donordata(x, type = NULL)
}
\arguments{
\item{x}{An object of class \code{individual}.}

\item{elements}{Requested list elements. Can be \code{"child"},
\code{"time"} or \code{"list"}. The default \code{"list"}
returns a list of both elements.}

\item{type}{Same as \code{elements}}
}
\value{
If \code{elements == "list"}, the function returns
a list of \code{"child"} and \code{"time"}. The function
returns a \code{"tibble"} for the choices \code{"child"} and
\code{"time"}.
}
\description{
This function returns the slots \code{"id"}, \code{"child"} and
\code{"time"}, either separate or as a \code{list} from
the \code{individual} object. The code
assumes that slots \code{"id"}, \code{"child"} and \code{"time"}
are up-to-data and consistent with the other slots.
}
\note{
\code{individual.to.donordata()} is deprecated, but exported
for legacy reasons
}
\examples{
require("donordata")
p <- donordata_to_individual(dnr = "smocc", id = 10001)
p
q <- individual_to_donordata(p)
q
}
\seealso{
\code{\link{xyz-class}}, \code{\link{bse-class}}
}
\author{
Stef van Buuren 2019
}
