% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualAN.R
\docType{class}
\name{individualAN-class}
\alias{individualAN-class}
\title{An S4 class to represent individual anthropometric data}
\description{
The \code{individualAN} class stores anthropometric measures as the collection of three \code{xyz}-class for height, weight and head circumference, respectively.
}
\section{Slots}{

\describe{
\item{\code{hgt}}{Length/height in cm (\code{xyz})}

\item{\code{wgt}}{Weight in kg (\code{xyz})}

\item{\code{hdc}}{Head circumference in cm (\code{xyz})}
}}
\examples{
# create object with height and weight measures
# add Z-scores calculate according to Dutch 1997 references
new("individualAN",
     hgt = new("xyz", yname = "hgt", x = c(0, 0.5), y = c(50, 70)),
     wgt = new("xyz", yname = "wgt", x = c(0, 0.3), y = c(3, 6)))

# calculate -2 and +2 centiles for height and head circumference
# using the WHO Child Growth Standard for girls
hgtref <- create.reference.call(libname = "who", prefix = "who2011",
                                sex = "female", yname = "hgt", sub = "")
hdcref <- create.reference.call(libname = "who", prefix = "who2011",
                                sex = "female", yname = "hdc", sub = "")
new("individualAN",
     hgt = new("xyz", yname = "hgt", x = c(0, 0, 0.25, 0.25), z = c(-2, 2, -2, 2), call = hgtref),
     wgt = new("xyz", yname = "wgt", x = c(0, 0, 1, 1), z = c(-2, 2, -2, 2), call = hdcref))
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{xyz-class}}
}

