% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualAN.R
\docType{class}
\name{individualAN-class}
\alias{individualAN-class}
\title{An S4 class to represent individual anthropometric data}
\description{
The \code{individualAN} class stores anthropometric measures as
the collection of \code{xyz}-class for height, weight,
head circumference, bmi, and weight for height.
}
\section{Slots}{

\describe{
\item{\code{hgt}}{Length/height in cm (\code{xyz})}

\item{\code{wgt}}{Weight in kg (\code{xyz})}

\item{\code{hdc}}{Head circumference in cm (\code{xyz})}

\item{\code{wfh}}{Weight for height kg/m (\code{xyz})}

\item{\code{bmi}}{Body mass index kg/m**2 (\code{xyz})}

\item{\code{dsc}}{D-score (D) (\code{xyz})}
}}

\examples{
# create object with height and weight measures
# do not calculate Z-scores since reference is undefined (no sex)
z <- new("individualAN",
     hgt = new("xyz", yname = "hgt", x = c(0, 0.5), y = c(50, 70), verbose = TRUE),
     wgt = new("xyz", yname = "wgt", x = c(0, 0.3), y = c(3, 6), verbose = TRUE))
data.frame(z)

# specify sex, and we get a reference
z <- new("individualAN",
     hgt = new("xyz", yname = "hgt", x = c(0, 0.5), y = c(50, 70), sex = "male"),
     wgt = new("xyz", yname = "wgt", x = c(0, 0.3), y = c(3, 6), sex = "male"))
data.frame(z)

# calculate -2 and +2 centiles for height and head circumference
# using the WHO Child Growth Standard for girls
hgtref <- "who_2006_hgt_female_"
hdcref <- "who_2007_hdc_female_"
z <- new("individualAN",
     hgt = new("xyz", yname = "hgt", x = c(0, 0, 0.25, 0.25),
       z = c(-2, 2, -2, 2), refcode = hgtref, pkg = "centile"),
     hdc = new("xyz", yname = "wgt", x = c(0, 0, 0.25, 0.25),
       z = c(-2, 2, -2, 2), refcode = hdcref, pkg = "centile"))
data.frame(z)
}
\seealso{
\code{\link{xyz-class}}
}
\author{
Stef van Buuren 2016-2020
}
