% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualBG.R
\docType{class}
\name{individualBG-class}
\alias{individualBG-class}
\title{An S4 class to represent individual background variables}
\description{
This class stores individual time-invariant background variables,
such as sex or birth weight. Each slot can only have
length zero (if entry is unset or missing) or length one.
}
\section{Slots}{

\describe{
\item{\code{sex}}{Either \code{"male"} or \code{"female"} (\code{character})}

\item{\code{etn}}{Etnicity code (\code{character})}

\item{\code{ga}}{Gestational age in weeks (\code{numeric})}

\item{\code{bw}}{Birth weight in grammes (\code{numeric})}

\item{\code{mult}}{Multiplicity, 1 = singleton, 2 = twin (\code{integer})}

\item{\code{goodhealth}}{In good health, \code{TRUE} or \code{FALSE} (\code{logical})}

\item{\code{hgtm}}{Height of mother in cm (\code{numeric})}

\item{\code{wgtm}}{Weight of mother in kg (\code{numeric})}

\item{\code{landm}}{Country of birth, mother (\code{character})}

\item{\code{edum}}{Education mother, \code{"low"}, \code{"middle"}, \code{"high"} (\code{character})}

\item{\code{agem}}{Mother age when pregnant, \code{"(15,25]"},
\code{"(25,36]"} or \code{"(36,45]"} (\code{character})}

\item{\code{smo}}{Mother smoked during pregnancy, \code{TRUE} or \code{FALSE} (\code{logical})}

\item{\code{hgtf}}{Height of father in cm (\code{numeric})}

\item{\code{wgtf}}{Weight of father in kg (\code{numeric})}

\item{\code{landf}}{Country of birth, father (\code{character})}

\item{\code{eduf}}{Education father, \code{"low"}, \code{"middle"}, \code{"high"} (\code{character})}
}}

