% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\docType{class}
\name{individualBS-class}
\alias{individualBS-class}
\title{An S4 class to represent individual broken stick estimates}
\description{
An S4 class to represent individual broken stick estimates
}
\section{Slots}{

\describe{
\item{\code{hgt}}{Length/height in cm (\code{bse})}

\item{\code{wgt}}{Weight in kg (\code{bse})}

\item{\code{hdc}}{Head circumference in cm (\code{bse})}
}}
\author{
Stef van Buuren 2016
}

