% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\docType{class}
\name{individualID-class}
\alias{individualID-class}
\title{An S4 class to represent individual identifiers}
\description{
An S4 class to represent individual identifiers
}
\section{Slots}{

\describe{
\item{\code{src}}{Name of data source (\code{character})}

\item{\code{id}}{Unique numerical individual identifier in source (\code{integer})}

\item{\code{name}}{Name of the individual (\code{character})}

\item{\code{dob}}{Date of birth of class (\code{date})}
}}
\examples{
# Create a new ID for Ron and Jasper
ron <- new("individualID", name = "Ron Smith",
           dob = as.character(as.Date("22-08-99", format = "\%d-\%m-\%y")),
           id = as.integer(204))
jasper <- new("individualID", name = "Jasper Fielding",
           id = as.integer(220))
}
\author{
Stef van Buuren 2016
}

