% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individual.R
\docType{class}
\name{individualID-class}
\alias{individualID-class}
\title{An S4 class to represent individual identifiers}
\description{
An S4 class to represent individual identifiers
}
\section{Slots}{

\describe{
\item{\code{src}}{Name of data source (\code{character})}

\item{\code{id}}{Unique numerical individual identifier in source (\code{integer})}

\item{\code{name}}{Name of the individual (\code{character})}

\item{\code{dob}}{Date of birth (\code{POSIXct})}
}}
\author{
Stef van Buuren 2016
}

