% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-individualID.R
\docType{class}
\name{individualID-class}
\alias{individualID-class}
\title{An S4 class to represent individual identifiers}
\description{
An S4 class to represent individual identifiers
}
\section{Slots}{

\describe{
\item{\code{id}}{Unique numerical individual identifier within \code{src} (\code{integer}, length 1)}

\item{\code{name}}{Name of the individual (\code{character})}

\item{\code{dob}}{Date of birth of class (\code{Date}, length 1)}

\item{\code{src}}{Study name (\code{character})}

\item{\code{dnr}}{Name of data object (\code{character})}
}}

\examples{
# Create a new ID for Ron and Jasper
ron <- new("individualID", name = c("Ron", "Smith"),
           dob = format(as.Date("1999-08-22"), "\%d-\%m-\%y"),
           id = as.integer(204))
jasper <- new("individualID", name = c("Jasper", "Fielding"),
           id = as.integer(220))
}
\author{
Stef van Buuren 2019
}
