% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cabinet.R
\docType{methods}
\name{[,cabinet,ANY,ANY,ANY-method}
\alias{[,cabinet,ANY,ANY,ANY-method}
\title{Extract part of cabinet}
\usage{
\S4method{[}{cabinet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An S4 object of class `cabinet`}

\item{i}{Indices specifying elements to extract.}

\item{j}{Not used}

\item{\dots}{Not used}

\item{drop}{Not used}
}
\description{
Extract part of cabinet
}
