% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bds_individual.R
\name{validate_bds_individual}
\alias{validate_bds_individual}
\title{Validate JSON file}
\usage{
validate_bds_individual(txt = NULL, verbose = TRUE,
  schema = "default")
}
\arguments{
\item{txt}{a JSON string, URL or file to be compared to the schema.}

\item{verbose}{include an error message when validation fails.}

\item{schema}{the schema to evaluate by. options are "default" and "string".}
}
\value{
A \code{boolean} with optional \code{errors} attribute.
}
\description{
This function takes data from a JSON source and compares it to a JSON schema
object. A \code{boolean} is returned indicating whether or not the provided
JSON data matches the schema.
}
\author{
Arjan Huizing 2019
}
