% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xyz.R
\docType{class}
\name{xyz-class}
\alias{xyz-class}
\title{Class for ages, measurements and standard deviation scores}
\description{
The \code{xyz} class stores three variables useful for growth curves:
\code{x} (usually age), \code{y} (height, weight, etc) and \code{z} (Z-score,
or standard deviation score (SDS)
corresponding to \code{y} given \code{x}). The class calculates
\code{z} (when \code{y} is specified), or \code{y} (when \code{z} is specified).
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Numeric vector of decimal ages (in years) at which
   measurements \code{y} are made.}
   \item{\code{y}:}{Numeric vector of measurement of length \code{length(x)}.}
   \item{\code{z}:}{Numeric vector of Z-score of length \code{length(x)}.}
   \item{\code{xname}:}{A character scalar with the name of the \code{x}
   variable,usually something like \code{"age"}.}
   \item{\code{yname}:}{A character scalar with the name of the measurement.
   Currently restricted to one of \code{c("hgt", "wgt", "hdc")} for height (in cm),
   weight (in kg) or head circumference (in cm), respectively.}
   \item{\code{zname}:}{A character scalar with the name of the \code{z}
   variable. By default, is it equal to \code{paste0(yname, ".z")}.}
   \item{\code{call}:}{An object of class \code{call} that specifies the
   appropriate reference table from the \pkg{clopus} package. Use the function
   \code{\link[clopus]{create.reference.call}} to set \code{call} manually.
   Alternatively, specify one or more of \code{libname}, \code{prefix},
   \code{sex},
   \code{yname} and \code{sub} arguments, which are passed down to
   \code{create.reference.call()}. See below for examples.}
}
}
\examples{
# specify three height measures
d <- new("xyz", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4))
# Note that Z-scores are added, relative to nl1997 reference for males
d
# Same, but now for a female
d <- new("xyz", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4), sex = "female")
d
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link[clopus]{create.reference.call}}
}
\keyword{classes}

