% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xyz.R
\docType{class}
\name{xyz-class}
\alias{xyz-class}
\title{An S4 class for anthropometric data}
\description{
The \code{xyz} class stores three variables useful for growth curves:
\code{x} (usually age), \code{y} (height, weight, etc) and \code{z} (Z-score,
or standard deviation score (SDS)
corresponding to \code{y} given \code{x}). The class calculates
\code{z} (when \code{y} is specified), or \code{y} (when \code{z} is specified).
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Numeric vector of decimal ages (in years) at which
   measurements \code{y} and Z-scores \code{z} are made. The length
   of \code{x} determines how many measurements (including \code{NA}'s)
   are stored.}
   \item{\code{y}:}{Numeric vector of measurement of length \code{length(x)}.}
   \item{\code{z}:}{Numeric vector of Z-score of length \code{length(x)}.}
   \item{\code{xname}:}{A character scalar with the name of the \code{x}
   variable,usually something like \code{"age"}.}
   \item{\code{yname}:}{A character scalar with the name of the measurement.
   Currently restricted to one of \code{c("hgt", "wgt", "hdc", "wfh", "bmi", "dsc")}
   for height (in cm), weight (in kg), head circumference (in cm),
   weight-for-height (kg/m), body mass index (kg/m^2) and D-score, respectively.}
   \item{\code{zname}:}{A character scalar with the name of the \code{z}
   variable. By default, is it equal to \code{paste0(yname, "_z")}.}
   \item{\code{call}:}{An object of class \code{call} that specifies the
   appropriate reference table from the \pkg{clopus} package. Use the function
   \code{\link[clopus]{create.reference.call}} to set \code{call} manually.
   Alternatively, specify one or more of \code{libname}, \code{prefix},
   \code{sex},
   \code{yname} and \code{sub} arguments, which are passed down to
   \code{create.reference.call()}. DEPRECATED.}
   \item{\code{refcode}:}{Since \code{minihealth 0.90.0}, Name of the reference. See
   \code{\link[centile]{make_refcode}}.}
   \item{\code{pkg}:}{Since \code{minihealth 0.90.0}, Name of the package containing
   the reference. See \code{\link[centile]{make_refcode}}. The default is
   \code{nlreferences}.}
   \item{\code{found}:}{Internal logical flag indicating
   whether the reference specified by
   \code{call} was actually found.}
   \item{\code{usetransform}:}{Logical indicating whether conversion to
   and from should be done using \code{clopus::transform_z()} and
   \code{clopus::transform_y()}.}
   \item{\code{transform}:}{Name of the \code{clopus} transform function.}
   \item{\code{sex}:}{Sex specified by user argument.}
   \item{\code{ga}:}{Gestational age specified by user argument.}
   \item{\code{age}:}{Decimal age specified by user argument.}
   \item{\code{digits}:}{Number of digits for \code{round()} function. Default
   is 3.}
}
}

\examples{
# specify length (in cm) for boy at ages 0, 0.2 and 0.5 years
d1 <- new("xyz", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4))
d1

# Z-scores of weight (in kg) at same ages
d2 <- new("xyz", x = c(0, 0.2, 0.5), y = c(3.2, 5.2, 7.0), yname = "wgt",
  sex = "female")
d2

# As as data.frame
data.frame(d2)

# Head circumference of girl, relative to WHO standard
d3 <- new("xyz", x = c(0, 0.2, 0.5), y = c(35, 38, 41), pkg = "centile",
refcode = "who_2007_hdc_female_")
d3
data.frame(d3)

# Standard weight centiles at age 0.5 year of Dutch girls
d5 <- new("xyz", x = rep(0.5, 5), z = -2:2, sex = "female", yname = "wgt")
d5

# calculate centiles at 1 year using the female WHO head circumference reference
d6 <- new("xyz", yname = "hdc", x = rep(1, 5), z = -2:2, pkg = "centile",
refcode = "who_2007_hdc_female_")
d6

# calculate P50 for preterms, female and born a GA week 32
d7 <- new("xyz", yname = "hgt", x = seq(0, 1, 1/12), z = rep(0, 13),
 refcode = "nl_2012_hgt_female_32")
d7

# use nlreferences::transform2y() as an alternative
d8 <- new("xyz", yname = "hgt", x = seq(0, 1, 1/12), z = rep(0, 13),
  usetransform = TRUE, ga = 32, sex = "female")
d8

# use nlreferences::transform2z function
d9 <- new("xyz", yname = "hgt", x = seq(0, 0.25, 1/12), y = c(43, 46, 48, 50),
  usetransform = TRUE, ga = 32, sex = "female", digits = 10)
d9
}
\author{
Stef van Buuren
}
\keyword{classes}
