% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xyz.R
\docType{class}
\name{xyz-class}
\alias{xyz-class}
\title{An S4 class for anthropometric data}
\description{
The \code{xyz} class stores three variables useful for growth curves:
\code{x} (usually age), \code{y} (height, weight, etc) and \code{z} (Z-score,
or standard deviation score (SDS)
corresponding to \code{y} given \code{x}). The class calculates
\code{z} (when \code{y} is specified), or \code{y} (when \code{z} is specified).
}
\details{
Changed Feb 2020: The function rounds `z` to 3 digits after the
decimal point.
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Numeric vector of decimal ages (in years) at which
   measurements \code{y} and Z-scores \code{z} are made. The length
   of \code{x} determines how many measurements (including \code{NA}'s)
   are stored.}
   \item{\code{y}:}{Numeric vector of measurement of length \code{length(x)}.}
   \item{\code{z}:}{Numeric vector of Z-score of length \code{length(x)}.}
   \item{\code{xname}:}{A character scalar with the name of the \code{x}
   variable,usually something like \code{"age"}.}
   \item{\code{yname}:}{A character scalar with the name of the measurement.
   Currently restricted to one of \code{c("hgt", "wgt", "hdc", "bmi", "dsc")}
   for height (in cm), weight (in kg), head circumference (in cm),
   body mass index and D-score, respectively.}
   \item{\code{zname}:}{A character scalar with the name of the \code{z}
   variable. By default, is it equal to \code{paste0(yname, ".z")}.}
   \item{\code{call}:}{An object of class \code{call} that specifies the
   appropriate reference table from the \pkg{clopus} package. Use the function
   \code{\link[clopus]{create.reference.call}} to set \code{call} manually.
   Alternatively, specify one or more of \code{libname}, \code{prefix},
   \code{sex},
   \code{yname} and \code{sub} arguments, which are passed down to
   \code{create.reference.call()}. See below for examples.}
   \item{\code{found}:}{Internal logical flag indicating
   whether the reference specified by
   \code{call} was actually found.}
   \item{\code{usetransform}:}{Logical indicating whether conversion to
   and from should be done using \code{clopus::transform_z()} and
   \code{clopus::transform_y()}.}
   \item{\code{transform}:}{Name of the \code{clopus} transform function.}
}
}

\examples{
# specify length (in cm) for boy at ages 0, 0.2 and 0.5 years
d1 <- new("xyz", x = c(0, 0.2, 0.5), y = c(51.0, 54.1, 63.4), sex = "male")
d1

# Z-scores of weight (in kg) at same ages
d2 <- new("xyz", x = c(0, 0.2, 0.5), y = c(3.2, 5.2, 7.0), yname = "wgt",
  sex = "female")
d2

# Obtain reference used to calculate Z-scores in d2
eval(d2@call)

# List available WHO references in clopus package
# find.references() only searches packages in the search list
# so let attach clopus first
library("clopus")
find.reference(libname = "who")
detach("package:clopus")

# Head circumference of girl, relative to WHO standard
d3 <- new("xyz", x = c(0, 0.2, 0.5), y = c(35, 38, 41),
libname = "clopus::who", prefix = "who2011", sex = "female", sub = "", yname = "hdc")
d3

# Shortcut specification of WHO standard
d3@call
d4 <- new("xyz", x = c(0, 0.2, 0.5), y = c(35, 38, 41), call = d3@call)
d4

# Standard weight centiles at age 0.5 year of Dutch girls
d5 <- new("xyz", x = rep(0.5, 5), z = -2:2, sex = "f", yname = "wgt")
d5

# calculate centiles at 1 year using the female WHO head circumference reference
ref <- clopus::create.reference.call(libname = "clopus::who", prefix = "who2011",
                                sex = "female", yname = "hdc", sub = "")
d6 <- new("xyz", yname = "hdc", x = rep(1, 5), z = -2:2, call = ref)
d6

# calculate P50 for preterms, female and born a GA week 32
ref <- clopus::create.reference.call(libname = "clopus::preterm", prefix = "pt2012a",
                                sex = "female", yname = "hgt", sub = "32")
d7 <- new("xyz", yname = "hgt", x = seq(0, 1, 1/12), z = rep(0, 13), call = ref)
d7

# use transform_y function instead of reference call
d8 <- new("xyz", yname = "hgt", x = seq(0, 1, 1/12), z = rep(0, 13),
  usetransform = TRUE, ga = 32, sex = "female")
d8

# use transform_z function
d9 <- new("xyz", yname = "hgt", x = seq(0, 0.25, 1/12), y = c(43, 46, 48, 50),
  usetransform = TRUE, ga = 32, sex = "female")
d9
}
\seealso{
\code{\link[clopus]{create.reference.call}}
}
\author{
Stef van Buuren 2016
}
\keyword{classes}
